pyproject.toml
[project]
name = "uni-ai-python"
version = "1.0.0"
description = "UniAI Knowledge Base Platform - Python AI Service"
requires-python = ">=3.11"
readme = "README.md"

dependencies = [
    # ---- Web Framework ----
    "fastapi>=0.115,<1.0",
    "uvicorn>=0.30,<1.0",
    "sse-starlette>=2.0,<3.0",

    # ---- RAG Framework ----
    "llama-index-core>=0.12,<0.13",
    "llama-index-vector-stores-milvus>=0.4,<0.5",

    # ---- Agent Orchestration ----
    "langgraph>=0.3,<0.4",
    "langchain-core>=0.3,<0.4",

    # ---- LLM ----
    "litellm==1.57.4",

    # ---- HuggingFace Compatibility ----
    "huggingface-hub>=0.23,<1.0",

    # ---- Document Parsing (light/base) ----
    "python-docx>=1.1,<2.0",
    "openpyxl>=3.1,<4.0",
    "xlrd>=2.0,<3.0",
    "beautifulsoup4>=4.12,<5.0",
    "readability-lxml>=0.8,<1.0",
    "markdown-it-py>=3.0,<4.0",
    "chardet>=5.2,<6.0",
    "PyMuPDF>=1.24,<2.0",
    "pandas>=2.2,<3.0",

    # ---- Vector Database / Storage ----
    "pymilvus>=2.4,<3.0",
    "asyncpg>=0.29,<1.0",
    "sqlalchemy[asyncio]>=2.0,<3.0",
    "elasticsearch[async]>=8.12,<9.0",
    "redis[hiredis]>=5.0,<6.0",
    "minio>=7.2,<8.0",
    "aio-pika>=9.4,<10.0",

    # ---- Task Queue ----
    "celery[redis]>=5.4,<6.0",

    # ---- Utilities ----
    "pydantic>=2.7,<3.0",
    "pydantic-settings>=2.3,<3.0",
    "pyyaml>=6.0,<7.0",
    "loguru>=0.7,<1.0",
    "httpx>=0.27,<1.0",
    "tenacity>=8.3,<9.0",
    "tiktoken>=0.7,<1.0",
    "prometheus-fastapi-instrumentator>=7.0,<8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0,<9.0",
    "pytest-asyncio>=0.23,<1.0",
    "pytest-cov>=5.0,<6.0",
    "ruff>=0.5,<1.0",
]

parse = [
    # Lightweight parse extras for file-type detection on Windows.
    # PDF parsing falls back to PyMuPDF by default.
    "python-magic-bin>=0.4.14; platform_system == 'Windows'",
]

parse-local = [
    "magic-pdf[full]>=0.10,<1.0",
    "marker-pdf>=1.0,<2.0",
    "python-magic-bin>=0.4.14; platform_system == 'Windows'",
]

ml = [
    "llama-index-embeddings-huggingface>=0.4,<0.5",
    "FlagEmbedding>=1.2,<2.0",
    "sentence-transformers>=3.0,<4.0",
    "huggingface-hub>=0.23,<1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["uni_ai_python"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.hatch.build.targets.wheel]
packages = ["uni_ai_python"]
